// src/types/channel.ts
export interface Channel {
  id: string;
  name: string;
  tagline: string;
  hls: string;
  namespace: string;
  is_live: string;
  is_movie: string;
  subtitle: string;
  image: string;
  jenis: string;
  premium: string;
  alpha_2_code: string;
  country_name: string;
  t_stamp: string;
  s_stamp: string;
  url_license: string;
  fake_event: string;
  header_iptv: string;
  header_license: string;
  category?: string;
}